Correções pós-code review

+++ src/routes/Users/Register.jsx
@@ -1,6 +1,6 @@
 import libraryFetch from '../../axios/config';
 import { useState } from 'react';
-import { useNavigate, useOutletContext } from 'react-router-dom';
+import { useNavigate, useOutletContext, Link } from 'react-router-dom';
 
 import './Login.css'
 
@@ -11,17 +11,20 @@ const Register = () => {
   const [username, setUsername] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('');
 
   const register = async (e) => {
     e.preventDefault();
+    if (password !== confirmPassword) {
+      console.error('Passwords do not match!');
+      return;
+    }
     try {
       const response = await libraryFetch.post('/register', {
         username: username,
         email: email,
         password: password
-      });
-
-      const { token } = response.data;
+      });const { token } = response.data;
 
       localStorage.setItem('token', token);
 
@@ -51,8 +54,13 @@ const Register = () => {
                 <label htmlFor='password'>Password</label>
                 <input name='password' type="password" id="password" onChange={(e) => setPassword(e.target.value)} />
             </div>
+            <div className="formControl">
+              <label htmlFor='confirmPassword'>Confirm Password</label>
+              <input name='confirmPassword' type="password" id="confirmPassword" onChange={(e) => setConfirmPassword(e.target.value)} />
+            </div>
             <input type="submit" value="Register" className="btn" />
         </form>
+        <p className='redirect'>Already have an account? <Link to={"/login"}>Login</Link></p>
     </div>
   )
 };
